services/maintenance-service/requirements.txt
asyncpg
psycopg2-binary
azure-functions
